requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
babel==2.16.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
courlan==1.3.2
dateparser==1.2.0
eval_type_backport==0.2.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
htmldate==1.9.2
httpcore==1.0.7
httplib2==0.22.0
httpx==0.27.2
idna==3.10
jsonpath-python==1.0.6
jusText==3.0.1
lxml==5.3.0
lxml_html_clean==0.4.1
mypy-extensions==1.0.0
numpy==2.1.3
pandas==2.2.3
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.2
pydantic_core==2.27.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
tld==0.13
tqdm==4.67.1
trafilatura==2.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.3
